requirements.txt
alembic==1.7.5
amqp==5.0.6
backports.entry-points-selectable==1.1.1
billiard==3.6.4.0
black==21.12b0
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==36.0.1
Deprecated==1.2.13
distlib==0.3.4
filelock==3.4.2
Flask==2.0.2
Flask-Admin==1.5.8
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
gevent==21.8.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
kombu==5.2.2
Mako==1.1.6
MarkupSafe==2.0.1
mypy-extensions==0.4.3
mysql-connector-python==8.0.28
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.0
prompt-toolkit==3.0.23
protobuf==3.19.1
pycparser==2.21
PyMySQL==1.0.2
pyparsing==3.0.6
python-dotenv==0.19.2
pytz==2021.3
redis==4.1.2
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.27
tomli==1.2.3
typing_extensions==4.0.1
urllib3==1.26.8
vine==5.0.0
virtualenv==20.10.0
virtualenv-clone==0.5.7
wcwidth==0.2.5
Werkzeug==2.0.2
wrapt==1.13.3
WTForms==3.0.0
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
